requirements.txt
absl-py==0.10.0
appdirs==1.4.4
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828496740/work
astroid @ file:///tmp/build/80754af9/astroid_1592495913683/work
attrs==19.3.0
backcall==0.2.0
black==19.10b0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1596809862272/work
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.3
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
entrypoints==0.3
flake8==3.8.3
flake8-bugbear==20.1.4
flake8-comprehensions==3.2.3
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1594901672528/work
geomloss==0.2.3
gitdb==4.0.5
GitPython==3.1.7
google-auth==1.21.1
google-auth-oauthlib==0.4.1
gql==0.2.0
graphql-core==1.1
grpcio==1.31.0
hausdorff==0.2.5
idna==2.10
imageio==2.9.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446433964/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206602906/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort @ file:///tmp/build/80754af9/isort_1598376199560/work
jedi==0.17.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1594732094290/work
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.34.0
lpips==0.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1594060262917/work
networkx==2.4
notebook @ file:///tmp/build/80754af9/notebook_1596838602091/work
numba==0.51.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233737064/work
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.42
packaging==20.4
pandas==1.1.5
pandocfilters==1.4.2
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pathspec==0.8.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1594307298074/work
portalocker==1.7.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes==2.2.0
Pygments==2.6.1
pylint @ file:///tmp/build/80754af9/pylint_1598624022510/work
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytorch3d==0.2.5
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
resample2d-cuda==0.0.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
sentry-sdk==0.16.5
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
subprocess32==3.5.4
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.8.13
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1596476591553/work
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.10
wandb==0.9.5
watchdog==0.10.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yacs==0.1.6
yapf==0.31.0
zipp==3.1.0
